♻️ [REFACTOR] #95 types/ 폴더 구조 개선
