--- src/types/gatherings/cancel.ts
@@ -1,5 +1,0 @@
-// src/types/gatherings/cancel.ts
-
-import { IGathering } from './models';
-
-export type CancelGatheringResponse = IGathering;
