--- src/types/gatherings/detail.ts
@@ -1,5 +1,0 @@
-// src/types/gatherings/detail.ts
-
-import { IGathering } from './models';
-
-export type GetGatheringDetailResponse = IGathering;
