+++ src/types/gatherings/index.ts
@@ -1,12 +1,5 @@
-// src/types/gatherings/index.ts
-
+// index.ts
 export * from './models';
-export * from './list';
-export * from './create';
-export * from './joined';
-export * from './detail';
-export * from './participants';
-export * from './cancel';
-export * from './join';
-export * from './leave';
+export * from './request';
+export * from './response';
 export * from './params';
