--- src/types/gatherings/join.ts
@@ -1,5 +1,0 @@
-// src/types/gatherings/join.ts
-
-export interface JoinGatheringResponse {
-  message: string;
-}
