--- src/types/gatherings/leave.ts
@@ -1,5 +1,0 @@
-// src/types/gatherings/leave.ts
-
-export interface LeaveGatheringResponse {
-  message: string;
-}
