--- src/types/gatherings/participants.ts
@@ -1,5 +1,0 @@
-// src/types/gatherings/participants.ts
-
-import { IParticipant } from './models';
-
-export type GetParticipantsResponse = IParticipant[];
