+++ src/types/gatherings/request.ts
@@ -1,7 +1,6 @@
-// src/types/gatherings/create.ts
-
 import { GatheringType, GatheringLocation } from '@/types/common';
-import { IGathering } from './models';
+
+// API 요청할 때 사용하는 Type
 
 export interface CreateGatheringRequest {
   location: GatheringLocation;
@@ -9,8 +8,6 @@ export interface CreateGatheringRequest {
   name: string;
   dateTime: string;
   capacity: number;
-  image?: File;
+  image?: File; // client-side only
   registrationEnd?: string;
 }
-
-export type CreateGatheringResponse = IGathering;
